Remove firebase dependency since we no longer use firebase
For some reason this package breaks in when being installed with yarn. This was never an issue before but recently netlify started defaulting to yarn instead of npm for codesandbox builds. This was causing students to suddenly be unable to deploy a live link of their projects. Removing the package fixes the issue